[CMLIB] Remove superfluous ULONG casts now that BinSize is ULONG.

+++ sdk/lib/cmlib/hivebin.c
@@ -24,7 +24,7 @@ HvpAddBin(
     PHCELL Block;
 
     BinSize = ROUND_UP(Size + sizeof(HBIN), HBLOCK_SIZE);
-    BlockCount = (ULONG)(BinSize / HBLOCK_SIZE);
+    BlockCount = BinSize / HBLOCK_SIZE;
 
     Bin = RegistryHive->Allocate(BinSize, TRUE, TAG_CM);
     if (Bin == NULL)
@@ -34,7 +34,7 @@ HvpAddBin(
     Bin->Signature = HV_HBIN_SIGNATURE;
     Bin->FileOffset = RegistryHive->Storage[Storage].Length *
                       HBLOCK_SIZE;
-    Bin->Size = (ULONG)BinSize;
+    Bin->Size = BinSize;
 
     /* Allocate new block list */
     OldBlockListSize = RegistryHive->Storage[Storage].Length;
